1. Модульные тесты [#505277 #528212]

+++ src/main/java/ru/job4j/converter/Converter.java
@@ -2,12 +2,12 @@ package ru.job4j.converter;
 
 public class Converter {
 
-    public static double rubleToEuro(double value) {
+    public static float rubleToEuro(float value) {
         // Assuming the conversion rate is 1 Euro = 70 Rubles
         return value / 70;
     }
 
-    public static double rubleToDollar(double value) {
+    public static float rubleToDollar(float value) {
         // Assuming the conversion rate is 1 Dollar = 60 Rubles
         return value / 60;
     }
